Update README.md
Stock Analysis With Excel VBA
1.	Overview of Project: Explain the purpose of this analysis.
The purpose of this Project is to apply refactoring code that is restructuring the codes already use in the same set of data without changing the end results. Refactoring data also helps the programmer simplify, easily comprehendable and use less memory. 
2.	Results: Using images and examples of your code, compare the stock performance between 2017 and 2018, as well as the execution times of the original script and the refactored script.
Based on the runtime results of the original script versus refactored script, we saved a tremendous amount of time in the refactored script averaging 0.125 seconds for both 2017 and 2018 years and still gets the same results. Other than the runtime the end results of the project did not change. We can further conclude that the data is not compromised
 
3.	Summary: In a summary statement, address the following questions.
a.	What are the advantages or disadvantages of refactoring code?

Advantages 
-	Reduce/Save time
-	Easy to read and follow
-	Cleaner outcome
-	Minimize redundancies/duplication
-	Obvious to other programmers 
-	Reduces and mitigate bugs

Disadvantages
-	Maintainability of team developers -personnel turnover reduces, it maybe challenging to the new programmer to trace back if something goes wrong if the author is no longer available 
-	Refactoring activities may generate architectural modifications that deteriorate the structural architecture of a software system – based on wiki
-	Code might break down as we minimize the process 

b.	How do these pros and cons apply to refactoring the original VBA script?
	As long as we obtain the same end result, we should consider refactoring codes more often as the pros outweighs the cons. Programmers must make sure that it’s easy to ready, clean and we organizer so that other programmers can easily follow the path.